requirements/common.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/

# Bleeding edge Django
Django>=1.7,<1.8
django-extensions>=1.4

# Configuration
django-environ>=0.3

# Forms
django-braces>=1.4

# Models
django-model-utils>=2.2

# API Rest
djangorestframework>=2.4
Markdown>=2.5
django-filter>=0.8

# Authentication
django-allauth>=0.18